Migrate to Qt6 and add release build script
- Port all source files from Qt4/Qt5 to Qt6 API changes
- Fix scatter file combo autoconnect (activated(QString) removed in Qt6)
- Fix SendReport moveToThread: create without parent, stop thread on teardown
- Add release.sh build script for Linux Qt6 builds

+++ Err/ErrorLookup.cpp
@@ -914,7 +914,7 @@ string ErrorLookup::FlashToolErrorMessage(FLASH_TOOL_RESULT ret) {
             prefix.append(buf);
     }
 
-    switch (ret) {
+    switch ((unsigned int)ret) {
         case FT_EXCEPTION:
             errMsg = "[SP Flash Tool] Common Error!";
             tipMsg = "SP Flash Tool common error.";
